+++ NEWS.md
@@ -1,3 +1,4 @@
 # strzip 0.0.0.9000
 
-* Added SMAZ 
+* Integrated SMAZ 
+* Integrated Unishox2